test interactions with mongodb utilizing the dependency 'mongodb-memory-server'

+++ application/services/databaseService.js
@@ -1,7 +1,8 @@
 const mongoose = require('mongoose');
 const {MongoMemoryServer} = require('mongodb-memory-server');
 
-const mongod = new MongoMemoryServer();
+let mongod;
+const TEST_ENVIRONMENT = 'test';
 
 module.exports.connectToMongoDatabase = async (databaseUrl) => {
     if (!databaseUrl) {
@@ -15,18 +16,27 @@ module.exports.connectToMongoDatabase = async (databaseUrl) => {
 }
 
 module.exports.connectToTestMongoDatabase = async () => {
-    const testDatabaseUrl = mongod.getUri();
+    mongod = await MongoMemoryServer.create();
 
-    await connectToMongoDatabase(testDatabaseUrl);
+    await mongoose.connect(mongod.getUri(), {
+        useNewUrlParser: true,
+        useUnifiedTopology: true
+    });
 }
 
 module.exports.closeDatabase = async () => {
+    if (process.env.NODE_ENV != TEST_ENVIRONMENT) {
+        throw Error('the database can ONLY be closed manually in test environments');
+    }
     await mongoose.connection.dropDatabase();
     await mongoose.connection.close();
     await mongod.stop();
 }
 
 module.exports.clearDatabase = async () => {
+    if (process.env.NODE_ENV != TEST_ENVIRONMENT) {
+        throw Error('the database can ONLY be cleared manually in test environments');
+    }
     const collections = mongoose.connection.collections;
     for (const key in collections) {
         const collection = collections[key];
